+++ martin_binance/client.py
@@ -4,7 +4,7 @@ gRPC async client for exchanges-wrapper
 __author__ = "Jerry Fedorenko"
 __copyright__ = "Copyright © 2021 Jerry Fedorenko aka VM"
 __license__ = "MIT"
-__version__ = "1.3.7.post2"
+__version__ = "1.4.0.b2"
 __maintainer__ = "Jerry Fedorenko"
 __contact__ = "https://github.com/DogsTailFarmer"
 
@@ -26,11 +26,12 @@ logger.addHandler(stream_handler)
 
 
 class Trade:
-    def __init__(self, channel_options, account_name, rate_limiter):
+    def __init__(self, channel_options, account_name, rate_limiter, symbol):
         self.channel = grpc.aio.insecure_channel(target='localhost:50051', options=channel_options)
         self.stub = api_pb2_grpc.MartinStub(self.channel)
         self.account_name = account_name
         self.rate_limiter = rate_limiter
+        self.symbol = symbol
         self.client: api_pb2.OpenClientConnectionId = None
         self.wait_connection = False
         self.trade_id = shortuuid.uuid()
@@ -46,8 +47,8 @@ class Trade:
             except UserWarning:
                 client = None
             else:
-                self.wait_connection = False
                 self.client = client
+                self.wait_connection = False
                 return True
 
     async def connect(self):
@@ -55,7 +56,9 @@ class Trade:
             _client = await self.stub.OpenClientConnection(api_pb2.OpenClientConnectionRequest(
                 trade_id=self.trade_id,
                 account_name=self.account_name,
-                rate_limiter=self.rate_limiter))
+                rate_limiter=self.rate_limiter,
+                symbol=self.symbol
+            ))
         except asyncio.CancelledError:
             pass  # Task cancellation should not be logged as an error.
         except grpc.RpcError as ex:
@@ -99,6 +102,10 @@ class Trade:
             raise
         else:
             # logger.info(f"send_request.res: {res}")
+            if res is None:
+                self.client = None
+                asyncio.create_task(self.get_client())
+                raise UserWarning("Can't get response, restart connection to gRPC server ...")
             return res
 
     async def for_request(self, _request, _request_type, **kwargs):